Add a rise-up maneuver
This complicated fiber-logic worked correctly on the first try.  I didn't expect that at all.  OK, there is a mild glitch on the client side; it thinks the player should move an extra pixel, and start to fall.  I'll sort that out later.  On the server side, though, this works just like I hoped.

Code is currently ugly and I'm sure it can be cleaned up.