Preparation for overhauling client/server sync.
Game now has a tick counter that starts at zero, and is incremented with every GameSpace update.  Tick count is sent to clients along with full registries, so clients can tell exactly how old the information is.  (Previously, the only possible action was to throw away all conflicting game state, and treat the new information as the current state of the world.)

Game now expects clients to send at_tick along with their requested moves.  Players aren't sending it yet, producing a bunch of warnings.  The idea is for players to specify that the action take place at a future time.  The Game stores all player moves until the right moment to process them.

Game now keeps a reference to the ServerPort, and consults it to find the ServerConnection corresponding to a Player's registry ID.  Player no longer keeps a reference to the Connection itself, because that was an unusual job for an Entity, and would have made planned changes harder.  Now we can freely copy the Player object without worrying about it holding onto objects besides its own state.

All input handling now resides in GameWindow.  Again, putting some of it into Player was a strange delegation of responsibility.

It's now possible to copy one GameSpace into another.  Not using that, just yet.

Player moves are now rebroadcast to other players.  Not making use of that, yet, either.