Mostly working again.  Soooooo many updates.
Full registry updates temporarily disabled.  I'll come back to that.  I wanted to see the deltas pulling their weight without the registry refreshes compensating for their inadequacies.

New delta, 'update_entities'.  Currently sent only when a block's HP is raised.  The 'build' move must be sent to the server, because only the server is allowed to generate registry IDs at the moment, and the 'build' move does that first.  Subsequent 'build's just increase the existing block's HP, but it's hard to pre-calculate that on the client side when we can't presuppose the existence of the block we'd be modifying.  So the server has to tell the client that the new block's HP has increased.

GameSpace no longer included as the first argument when creating an entity.  That was getting too confusing.  Now, the field is set when 'space << entity' happens.  This leads to a couple other wrinkles:

1) On the client side, it's hard to assign an entity's owner, based on the registry ID, before we know which space's registry to look up that ID.  New OwnedEntity base class handles this, with a hook Entity.on_added_to_space (called by space=) that OwnedEntity overrides.  Block and Pellet now extend OwnedEntity.

2) Entity.warp() is called during update_from_json() before the object is injected into a space.  It now carefully uses space.process_moving_entity() only *after* it has been injected.

Code in ClientEngine for processing player moves was redundant; apply_deltas() handles that now.

Late updates received by the server are now delayed until the *next* tick.

Bugs fixed in grid creation and at().  It was "working", but only by accident.  It now makes more sense to examine a grid in a debugger.

Conflict checking is now done in 'space << entity'.

New players are added into the actual, correct center of the space now.