Re-enable registry syncs
Better handling of the case where the client falls way behind the server: Just start over with the server's conception of the world.  There is nothing to be gained by presenting a smoother, but inaccurate, picture of how things are, if we're slipping further and further behind.

Slightly better formula for constraining the server to 60 updates per second.  This version of Gosu apparently produces 60 frames per second, not 62-ish, so we can drop the client-side timing adjustments.

Client and server both now have options to enable profiling output and debug network traffic.