Moving toward testability
I think the only way I'll nail down the client/server interactions, to the point where they really stay solidly in sync, is to write some real tests.  So I've laid the groundwork for that.  I have a spec that wires up a client-space and server-space in the same VM, using fake objects for the ENet connection, so I can get complete visibility and control over the interactions.  It's also now possible to check whether one space == another space, and get a reasonable answer.

Discovered that it's a bad idea to have Cell extend Set.  Set stores everything as hash keys, but Entities are mutable.  Hash keys should never mutate while in use.

Also decided it's not advisable to have entities directly storing references to each other, as it complicates the logic on the client side.  Better to store only the other object's registry ID, and look up the actual object from the current GameSpace when needed.