More consistency improvements
When the server sends a complete registry, we don't want to update it; we want to display it exactly as-is.  The *next* frame should be an updated copy of this one.

When the server sends add_npcs, it's saying these NPCs were added *during* this frame.  They aren't updated during the same tick.  So, on the client-side, wait until after the update to insert the new objects.

Make Player equality work.  Behavior of == and <=> must be consistent, so we should use all_state array for both.  Put player_name first so we sort based on that.

Make it possible to create Blocks of different levels.