A whole lot more revisions.
Hard work has paid off.  "Rise up" move now looks rock-solid.  RSpec test proves client and server stay exactly in sync, at least when they run at the same rate.

Added hash.rb with a short'n'sweet utility function, for recursively turning all the hash-keys into symbols.  No longer necessary to figure out when the code is expecting a string for a key.  We just fix 'em all when the hash is recovered from JSON.  Used for messages received over the network, and for state restored from storage.

It was getting tricky to keep the client in sync with the server, when they followed different logic.  The right solution seemed to be to allow the client to do anything the server did, including create new entities.  For that, the entity IDs needed to be predictable.  So we now use increasing integers for IDs, instead of UUIDs.  The space tracks which ID we should dispense next, and this information is conveyed to the client.

Likewise, when generating a move, the client creates its own delta and applies it, so it will match what the server (and other clients) will do.

Using a fiber for "rise up" wasn't working, because the server couldn't sync the client to any point in the middle of the move.  Replaced this with a ComplexMove, an object with state which is tied to the player executing it.  ComplexMoves are sent to the client during a registry broadcast.

Factored out Serializable, a module for handling the JSON <-> object conversion.  It's now used by ComplexMove as well as Entity.

Likewise, constants from Entity now live in EntityConstants, and are used by RiseUp.

Worlds now have a name and an ID.