Groundwork for simple authentication
Client and server can now encrypt sensitive information when necessary.  Uses Diffie-Hellman key exchange to establish the symmetric encryption key, with a default key-size of 1024 bits.  New client-side option 'key-size' to customize this, because generating a 1024-bit key takes too long for repeated runs.

Client now sends a password after handshake.  Encryption is used to protect the password in transit.  If the password is wrong, the server kicks the client off.  Currently the password is hard-coded, so fixing that (adding player persistence) is my next task.

DH takes long enough that it's worth telling the player the reason for the delay.  We now pop up a message in the center of the screen.

Menu generation has been broken up into separate methods, making for cleaner code.

GameWindow constructor now accepts an options-hash, like Game.

Fixed some hard-coded references to the font size.