Introducing player persistence
All we store so far is the password hash.  If the player isn't known, it is created and saved.  No other password management facilities yet (change password, reset password).

Password is hashed using crypt(), client-side, before encryption.  The server doesn't care.  It just compares the password hash it is given with the one it previously stored.  But this does mean that someone who gets to peek at the player storage file will have to actually break the crypt()ed passwords before being to impersonate someone.