Bug fixes, start of grab-block feature.
Fixed a bug where the player could use a build-block to walk away from solid ground.  The build-block is not to be used as solid ground for the player building it.  I forbid this.

Fixed a bug with the initial update -- server was sending an incomplete frame to the new player.

Fixed the class methods in Entity.

Fixed the registry broadcast, so it no longer sends to players who aren't yet logged in.

Fixed specs so they pass again.  Had to be adjusted after the addition of password auth.

I grew tired of the combinatorial explosion, created by telling every entity-type whether it was opaque or transparent to every other type.  Extracted that logic out into a single module, Transparency.

Building is now done by holding Shift and right-clicking.  Unmodified right-click now "grabs" an entity and moves it around, or releases it.  This isn't working properly yet.

Introduced a teleporter, and a destination for it.  This isn't working yet either, as I don't have a reasonable way to tie one to the other.  My plan is to get the "grab" feature working first, then automatically create-and-grab a teleporter-destination when a teleporter is created.