Better block-moving (I think)
Instead of sending everything the client knows about the object, it sends only the velocity to apply.  update_from_json() implementations now meet the contract that they accept partial information, and if fields are missing, it leaves them unchanged.  The velocity the client sends is now 1/6 of the delta, intended to compensate for the fact that motions are scheduled for 6 ticks in the future.  Interestingly, this produces a behavior where the grabbed block sort of bounces around the mouse, converging on its position fairly quickly, but with some wobble that I rather like.