Teleporters work now -- fun!
Teleporter is the owner of a Destination.  When an object intersects with the teleporter's center, and the destination is free of obstructions (objects opaque to the teleportee), its location is changed.  Its velocity is left alone.

Creating teleporter destinations is pretty much as I planned: You create the teleporter, and it automatically creates a destination too.  Ended up doing this by adding a hook, :on_create, to the local (ClientEngine's) copy of the add_npcs delta.  As soon as the delta has been executed, thus creating the teleporter and assigning it an ID, the hook (a Proc) is invoked and passed the teleporter as an argument.  The hook can then create a Destination whose owner is the new Teleporter.  There is still a bug here, because occasionally the Destination doesn't get created -- perhaps because the mouse moves during that 1/10 second.  The Destination should really be created exactly where the teleporter is.  Anyway, the Destination has a hook too, causing it to be grabbed right after it's created, as if right-clicked.

Needed a convenient way to find objects owned by another.  GameSpace now tracks this for us.  It has to be the GameSpace, because we keep copying it in the ClientEngine.  Try to put ownership information somewhere else, and your list of owned items quickly fills up with duplicates.

Ditched the flag in deserialization that was trying to predict whether the registry ID was supposed to be included in the JSON, or generated when the object is added to the space.  Logic for this got unduly complicated.