Make snap-to-grid work
With the way grabbed-entities wobble around, even with snap turned on, they weren't ending up in the right place.  Tried a couple things using 'update_entities', but it wasn't working.  So I decided it would be easiest to tell the game exactly what I wanted: actually snap the entity to the grid.  Works like a charm now.

Since I haven't yet found a use for the display of player-score, I replaced it with some diagnostic information about the entities under the mouse.