Create Base
Doesn't actually function, yet.  But it can stick to walls or ceilings, just like a player.  It's transparent to players.  And if it falls, it changes its orientation... but unlike a player, it always turns with its "feet" towards the direction it's going, giving it a chance to stick to a wall or ceiling.