Players can die now
Players have HP.  Just one for now, to ease testing.  If damaged, a player becomes a Ghost.  The Gecko object is destroyed in favor of the Ghost, and everyone is notified of the change of circumstance.

Likewise, Ghost players can now respawn.  A Ghost who clicks on the screen automatically selects the closest available Base to the click, and moves there rapidly.  Assuming it's still unoccupied on arrival, the Ghost is discarded in favor of a shiny new Gecko.

Not coincidentally, I'm now depending less on the player's registry ID (because that can now change) to identify a player, and more on the player name (because that cannot).