Fix a bug with die/respawn
Occasionally the GameClient.player_id field wouldn't get updated during a death or a respawn.  This was apparently due to the server sending a full registry update at just the right moment, so that the client misses the transition that would usually be handled by Move::Spawn or Player#die().  We could have the server send the client another you_are message, but it's not necessary.  If the server is sending us a Player object we don't have, and its player_name is ours, then it must be our player entity.