Add Slime
Slime moves like a lemming.  It's also cute, like a lemming.  It's not as fragile as a lemming; in fact it tends to do a lot of damage to whatever it touches.  Because it's made of acid.  (Cute acid.)  It does more damage when it falls on things, but it also damages what it walks on, and what it bumps into.  When it's trapped, it does damage to its prison... slowly.

It's now possible to rotate the selected (pointed-at) or grabbed entity, using the [ and ] keys.  Useful for positioning bases on walls/ceilings.

Fixed a bug where all entities would be woken up, client-side, simply because the << operator was being used to add them to the GameSpace.  This shouldn't be done for objects sent by the server, such as during full registry updates.

Fixed a bug where a Gecko dying while on a wall would become a sideways-ghost, which just looked silly.