Smarter brick and cement
Brick and cement blocks now fall unless they line up *exactly* with supporting objects on their sides.  Additional logic ensures that a falling block will stop at the right level, even at high velocity, if it encounters a source of support.  Getting this right was tricky, so -- unit test.

No change to the behavior of dirt or steel.  For now I'm happy with them as they are, but I might make a similar change to steel blocks at some point, for consistency.