Make Slime fall through one-cell-wide holes
Prior to this change, they almost never did.  Their velocity would carry them from block A to block C, never noticing that block B was missing.  Factored out a reusable slide_around() method from the Gecko code.  Now using it to ensure that a Slime encountering a big-enough hole always falls into it.

There's a potential problem with this solution, since it takes the leftover horizontal velocity and applies it vertically.  If something causes a Slime to move very quickly along the floor, this will produce a noticeable sudden drop.  It's not noticeable now because Slime generally move at a sedate pace.