Make the player's death violent
Now, the Gecko is replaced with two entities upon destruction - the Ghost, and a Nuke.  The Nuke flares up and fizzles out, then disappears.  It can't be destroyed earlier than scheduled.  It behaves a lot like a dirt block in terms of motion (affected by gravity, falls if nothing under it), so a player who dies in a black hole may produce a Nuke that swirls around in the hole.

The Nuke also does several points of damage to all entities within a certain range of it.  If you're trying to build something, this imposes an added cost: your beautiful creation is damaged.  If you were killed by something touching you (a Slime, say), you might avenge your own death.