Use a VM to run programs
Previously, a simple 'while true' loop would freeze up the server and client.  Now each program is compiled to a list of simple instructions, and runs in a little VM, a bit at a time.  Each instruction has a cost, and when the available CPU is used up, the program goes on hold until the next tick.